refactor(k-sigma anomaly): k value is a float

+++ src/sdk/python/rtdip_sdk/pipelines/data_wranglers/spark/data_quality/k_sigma_anomaly_detection.py
@@ -30,7 +30,7 @@ class KSigmaAnomalyDetection(WranglerBaseInterface):
         spark: SparkSession,
         df: DataFrame,
         column_names: list[str],
-        k_value: int = 3,
+        k_value: float = 3.0,
         use_median: bool = False,
     ) -> None:
         if len(column_names) == 0:
